cap-04/DIP: Adiciona a inversao de dependencia da classe Cotuba, passando a depender das interfaces estaveis GeradorPDF, GeradorEPUB e RenderizadorMDParaHTML.
