+++ src/main/java/cotuba/application/Cotuba.java
@@ -1,10 +1,13 @@
 package cotuba.application;
 
-import cotuba.epub.GeradorEPUB;
-import cotuba.pdf.GeradorPDF;
-import cotuba.md.RenderizadorMDParaHTML;
 import cotuba.domain.Capitulo;
 import cotuba.domain.Ebook;
+import cotuba.epub.GeradorEPUB;
+import cotuba.epub.GeradorEPUBImpl;
+import cotuba.md.RenderizadorMDParaHTML;
+import cotuba.md.RenderizadorMDParaHTMLImpl;
+import cotuba.pdf.GeradorPDF;
+import cotuba.pdf.GeradorPDFImpl;
 
 import java.nio.file.Path;
 import java.util.List;
@@ -12,7 +15,7 @@ import java.util.List;
 public class Cotuba {
 
   public void executa(final String formato, final Path diretorioDosMD, final Path arquivoDeSaida) {
-    final RenderizadorMDParaHTML renderizador = new RenderizadorMDParaHTML();
+    final RenderizadorMDParaHTML renderizador = new RenderizadorMDParaHTMLImpl();
 
     final List<Capitulo> capitulos = renderizador.renderiza(diretorioDosMD);
 
@@ -22,11 +25,11 @@ public class Cotuba {
     ebook.setCapitulos(capitulos);
 
     if ("pdf".equals(formato)) {
-      final GeradorPDF geradorPDF = new GeradorPDF();
+      final GeradorPDF geradorPDF = new GeradorPDFImpl();
 
       geradorPDF.gera(ebook);
     } else if ("epub".equals(formato)) {
-      final GeradorEPUB geradorEPUB = new GeradorEPUB();
+      final GeradorEPUB geradorEPUB = new GeradorEPUBImpl();
 
       geradorEPUB.gera(ebook);
     } else {
